Re-organize HA discovery messages
