+++ src/integrations/home_assistant/discovery.py
@@ -2,86 +2,31 @@ from __future__ import annotations
 
 import json
 import logging
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING, Any, override
 
-import inflection
 from saic_ismart_client_ng.api.vehicle_charging import (
     ChargeCurrentLimitCode,
     ScheduledChargingMode,
 )
 
 import mqtt_topics
+from integrations.home_assistant.availability import (
+    HaCustomAvailabilityConfig,
+    HaCustomAvailabilityEntry,
+)
+from integrations.home_assistant.base import HomeAssistantDiscoveryBase
+from integrations.home_assistant.utils import decode_as_utf8, snake_case
 from publisher.mqtt_publisher import MqttPublisher
 from vehicle import RefreshMode, VehicleState
 
 if TYPE_CHECKING:
-    from collections.abc import Collection
-
     from configuration import Configuration
     from vehicle_info import VehicleInfo
 
 LOG = logging.getLogger(__name__)
 
 
-class HaCustomAvailabilityEntry:
-    def __init__(
-        self,
-        *,
-        topic: str,
-        template: str | None = None,
-        payload_available: str = "online",
-        payload_not_available: str = "offline",
-    ) -> None:
-        self.__topic = topic
-        self.__template = template
-        self.__payload_available = payload_available
-        self.__payload_not_available = payload_not_available
-
-    def to_dict(self) -> dict[str, Any]:
-        result = {
-            "topic": self.__topic,
-            "payload_available": self.__payload_available,
-            "payload_not_available": self.__payload_not_available,
-        }
-        if self.__template:
-            result.update({"value_template": self.__template})
-        return result
-
-    def __key(self) -> Collection[str | None]:
-        return (
-            self.__topic,
-            self.__template,
-            self.__payload_available,
-            self.__payload_not_available,
-        )
-
-    def __hash__(self) -> int:
-        return hash(self.__key())
-
-    def __eq__(self, other: object) -> bool:
-        if isinstance(other, HaCustomAvailabilityEntry):
-            return self.__key() == other.__key()
-        return NotImplemented
-
-
-class HaCustomAvailabilityConfig:
-    def __init__(
-        self,
-        *,
-        rules: list[HaCustomAvailabilityEntry],
-        mode: str = "all",
-    ) -> None:
-        self.__rules = rules
-        self.__mode = mode
-
-    def to_dict(self) -> dict[str, Any]:
-        return {
-            "availability": [r.to_dict() for r in set(self.__rules)],
-            "availability_mode": self.__mode,
-        }
-
-
-class HomeAssistantDiscovery:
+class HomeAssistantDiscovery(HomeAssistantDiscoveryBase):
     def __init__(
         self,
         vehicle_state: VehicleState,
@@ -128,253 +73,130 @@ class HomeAssistantDiscovery:
         LOG.debug("Publishing Home Assistant discovery messages")
 
         # Gateway Control
-        self.__publish_select(
-            mqtt_topics.REFRESH_MODE,
-            "Gateway refresh mode",
-            [m.value for m in RefreshMode],
-            entity_category="config",
-            icon="mdi:refresh",
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_number(
-            mqtt_topics.REFRESH_PERIOD_ACTIVE,
-            "Gateway active refresh period",
-            entity_category="config",
-            unit_of_measurement="s",
-            icon="mdi:timer",
-            min_value=30,
-            max_value=60 * 60,
-            step=1,
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_number(
-            mqtt_topics.REFRESH_PERIOD_INACTIVE,
-            "Gateway inactive refresh period",
-            entity_category="config",
-            unit_of_measurement="s",
-            icon="mdi:timer",
-            min_value=1 * 60 * 60,
-            max_value=5 * 24 * 60 * 60,
-            step=1,
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_number(
-            mqtt_topics.REFRESH_PERIOD_AFTER_SHUTDOWN,
-            "Gateway refresh period after car shutdown",
-            entity_category="config",
-            unit_of_measurement="s",
-            icon="mdi:timer",
-            min_value=30,
-            max_value=12 * 60 * 60,
-            step=1,
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_number(
-            mqtt_topics.REFRESH_PERIOD_INACTIVE_GRACE,
-            "Gateway grace period after car shutdown",
-            entity_category="config",
-            unit_of_measurement="s",
-            icon="mdi:timer",
-            min_value=30,
-            max_value=12 * 60 * 60,
-            step=1,
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_sensor(
-            mqtt_topics.REFRESH_PERIOD_CHARGING,
-            "Gateway charging refresh period",
-            entity_category="diagnostic",
-            unit_of_measurement="s",
-            icon="mdi:timer",
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_sensor(
-            mqtt_topics.REFRESH_PERIOD_ERROR,
-            "Gateway error refresh period",
-            entity_category="diagnostic",
-            unit_of_measurement="s",
-            icon="mdi:timer",
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_sensor(
-            mqtt_topics.REFRESH_LAST_ACTIVITY,
-            "Last car activity",
-            device_class="timestamp",
-            entity_category="diagnostic",
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_sensor(
-            mqtt_topics.REFRESH_LAST_CHARGE_STATE,
-            "Last charge state",
-            device_class="timestamp",
-            entity_category="diagnostic",
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_sensor(
-            mqtt_topics.REFRESH_LAST_VEHICLE_STATE,
-            "Last vehicle state",
-            device_class="timestamp",
-            entity_category="diagnostic",
-            custom_availability=self.__system_availability_config,
-        )
-        self.__publish_sensor(
-            mqtt_topics.REFRESH_LAST_ERROR,
-            "Last poll error",
-            device_class="timestamp",
-            entity_category="diagnostic",
-            custom_availability=self.__system_availability_config,
-        )
+        self.__publish_gateway_sensors()
 
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.INFO_LAST_MESSAGE_CONTENT,
             "Last car message",
-            entity_category="diagnostic",
             enabled=False,
+            entity_category="diagnostic",
             custom_availability=self.__system_availability_config,
         )
 
+        self.__publish_climate_sensors()
+
         # Complex sensors
-        self.__publish_remote_ac()
-        self.__publish_heated_seats()
-        self.__publish_vehicle_tracker()
         self.__publish_scheduled_charging()
         self.__publish_scheduled_battery_heating()
 
         # Switches
-        self.__publish_switch(mqtt_topics.DRIVETRAIN_CHARGING, "Charging")
-        self.__publish_switch(
+        self._publish_switch(mqtt_topics.DRIVETRAIN_CHARGING, "Charging")
+        self._publish_switch(
             mqtt_topics.DRIVETRAIN_BATTERY_HEATING,
             "Battery heating",
             icon="mdi:heat-wave",
         )
-        self.__publish_switch(mqtt_topics.WINDOWS_DRIVER, "Window driver")
-        self.__publish_switch(mqtt_topics.WINDOWS_PASSENGER, "Window passenger")
-        self.__publish_switch(mqtt_topics.WINDOWS_REAR_LEFT, "Window rear left")
-        self.__publish_switch(mqtt_topics.WINDOWS_REAR_RIGHT, "Window rear right")
-
-        if self.__vin_info.has_sunroof:
-            self.__publish_switch(mqtt_topics.WINDOWS_SUN_ROOF, "Sun roof")
-            self.__publish_binary_sensor(mqtt_topics.WINDOWS_SUN_ROOF, "Sun roof")
-        else:
-            self.__unpublish_ha_discovery_message("switch", "Sun roof")
-            self.__unpublish_ha_discovery_message("binary_sensor", "Sun roof")
-
-        self.__publish_switch(
-            mqtt_topics.CLIMATE_BACK_WINDOW_HEAT,
-            "Rear window defroster heating",
-            icon="mdi:car-defrost-rear",
-            payload_on="on",
-            payload_off="off",
-        )
-        self.__publish_switch(
-            mqtt_topics.LOCATION_FIND_MY_CAR,
-            "Find my car",
-            icon="mdi:car-search",
-            payload_on="activate",
-            payload_off="stop",
-        )
+        self.__publish_windows_sensors()
 
         # Locks
-        self.__publish_lock(
-            mqtt_topics.DOORS_LOCKED, "Doors Lock", icon="mdi:car-door-lock"
-        )
-        self.__publish_lock(
-            mqtt_topics.DOORS_BOOT,
-            "Boot Lock",
-            icon="mdi:car-door-lock",
-            state_locked="False",
-            state_unlocked="True",
-        )
-        self.__publish_lock(
+        self.__publish_doors_sensors()
+
+        self._publish_lock(
             mqtt_topics.DRIVETRAIN_CHARGING_CABLE_LOCK,
             "Charging Cable Lock",
             icon="mdi:lock",
         )
 
         # Target SoC
-        self.__publish_number(
+        self._publish_number(
             mqtt_topics.DRIVETRAIN_SOC_TARGET,
             "Target SoC",
+            enabled=self.__vin_info.supports_target_soc,
             device_class="battery",
             unit_of_measurement="%",
+            icon="mdi:battery-charging-70",
+            mode="slider",
             min_value=40,
             max_value=100,
             step=10,
-            mode="slider",
-            icon="mdi:battery-charging-70",
-            enabled=self.__vin_info.supports_target_soc,
         )
-        self.__publish_select(
+        options = [
+            m.limit
+            for m in ChargeCurrentLimitCode
+            if m != ChargeCurrentLimitCode.C_IGNORE
+        ]
+        self._publish_select(
             mqtt_topics.DRIVETRAIN_CHARGECURRENT_LIMIT,
             "Charge current limit",
-            [
-                m.limit
-                for m in ChargeCurrentLimitCode
-                if m != ChargeCurrentLimitCode.C_IGNORE
-            ],
+            options,
+            entity_category=None,
+            enabled=True,
+            value_template="{{ value }}",
+            command_template="{{ value }}",
             icon="mdi:current-ac",
+            custom_availability=None,
         )
 
         # Standard sensors
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_SOC,
             "SoC",
             device_class="battery",
             state_class="measurement",
             unit_of_measurement="%",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_SOC_KWH,
             "SoC_kWh",
             device_class="ENERGY_STORAGE",
             state_class="measurement",
-            icon="mdi:battery-charging-70",
             unit_of_measurement="kWh",
+            icon="mdi:battery-charging-70",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_TOTAL_BATTERY_CAPACITY,
             "Total Battery Capacity",
+            entity_category="diagnostic",
             device_class="ENERGY_STORAGE",
             state_class="measurement",
-            icon="mdi:battery-high",
             unit_of_measurement="kWh",
-            entity_category="diagnostic",
+            icon="mdi:battery-high",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_LAST_CHARGE_ENDING_POWER,
             "Last Charge SoC kWh",
             device_class="ENERGY_STORAGE",
             state_class="measurement",
-            icon="mdi:battery-charging-70",
             unit_of_measurement="kWh",
+            icon="mdi:battery-charging-70",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_POWER_USAGE_SINCE_LAST_CHARGE,
             "Energy Usage Since Last Charge",
+            enabled=False,
             device_class="ENERGY_STORAGE",
             state_class="measurement",
-            icon="mdi:battery-charging-70",
             unit_of_measurement="kWh",
-            enabled=False,
+            icon="mdi:battery-charging-70",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_POWER_USAGE_OF_DAY,
             "Energy Usage of the Day",
+            enabled=False,
             device_class="ENERGY_STORAGE",
             state_class="measurement",
-            icon="mdi:battery-charging-70",
             unit_of_measurement="kWh",
-            enabled=False,
+            icon="mdi:battery-charging-70",
         )
 
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_REMAINING_CHARGING_TIME,
             "Remaining charging time",
             device_class="duration",
             state_class="measurement",
             unit_of_measurement="s",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_REMAINING_CHARGING_TIME,
             "Charging finished",
             device_class="timestamp",
@@ -392,12 +214,12 @@ class HomeAssistantDiscovery:
                 ]
             ),
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CHARGING_LAST_START,
             "Last Charge Start Time",
             device_class="timestamp",
-            value_template="{{ value | int | timestamp_utc }}",
             icon="mdi:clock-start",
+            value_template="{{ value | int | timestamp_utc }}",
             custom_availability=HaCustomAvailabilityConfig(
                 rules=[
                     self.__system_availability,
@@ -411,12 +233,12 @@ class HomeAssistantDiscovery:
                 ]
             ),
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CHARGING_LAST_END,
             "Last Charge End Time",
             device_class="timestamp",
-            value_template="{{ value | int | timestamp_utc }}",
             icon="mdi:clock-end",
+            value_template="{{ value | int | timestamp_utc }}",
             custom_availability=HaCustomAvailabilityConfig(
                 rules=[
                     self.__system_availability,
@@ -430,57 +252,57 @@ class HomeAssistantDiscovery:
                 ]
             ),
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CHARGING_TYPE,
             "Charging Mode",
-            entity_category="diagnostic",
             enabled=False,
+            entity_category="diagnostic",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.BMS_CHARGE_STATUS,
             "BMS Charge Status",
-            entity_category="diagnostic",
             enabled=False,
+            entity_category="diagnostic",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_MILEAGE,
             "Mileage",
             device_class="distance",
             state_class="total_increasing",
             unit_of_measurement="km",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_MILEAGE_OF_DAY,
             "Mileage of the day",
+            enabled=False,
             device_class="distance",
             state_class="total_increasing",
             unit_of_measurement="km",
-            enabled=False,
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_MILEAGE_SINCE_LAST_CHARGE,
             "Mileage since last charge",
+            enabled=False,
             device_class="distance",
             state_class="total_increasing",
             unit_of_measurement="km",
-            enabled=False,
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CURRENT_JOURNEY,
             "Mileage of journey",
+            enabled=False,
             device_class="distance",
             state_class="total_increasing",
             unit_of_measurement="km",
             value_template='{{ value_json["distance"] | int(0) }}',
-            enabled=False,
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CURRENT_JOURNEY,
             "Identifier of journey",
-            value_template='{{ value_json["id"] | int(0) }}',
             enabled=False,
+            value_template='{{ value_json["id"] | int(0) }}',
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_AUXILIARY_BATTERY_VOLTAGE,
             "Auxiliary battery voltage",
             device_class="voltage",
@@ -488,42 +310,42 @@ class HomeAssistantDiscovery:
             unit_of_measurement="V",
             icon="mdi:car-battery",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_RANGE,
             "Range",
             device_class="distance",
             unit_of_measurement="km",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_FOSSIL_FUEL_RANGE,
             "Fossil fuel range",
+            enabled=self.__vin_info.has_fossil_fuel,
             device_class="distance",
             unit_of_measurement="km",
-            enabled=self.__vin_info.has_fossil_fuel,
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_FOSSIL_FUEL_PERCENTAGE,
             "Fossil fuel percentage",
+            enabled=self.__vin_info.has_fossil_fuel,
             state_class="measurement",
             unit_of_measurement="%",
             icon="mdi:fuel",
-            enabled=self.__vin_info.has_fossil_fuel,
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CURRENT,
             "Current",
             device_class="current",
             state_class="measurement",
             unit_of_measurement="A",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_VOLTAGE,
             "Voltage",
             device_class="voltage",
             state_class="measurement",
             unit_of_measurement="V",
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.DRIVETRAIN_POWER,
             "Power",
             device_class="power",
@@ -531,7 +353,108 @@ class HomeAssistantDiscovery:
             unit_of_measurement="kW",
         )
 
-        self.__publish_sensor(
+        self.__publish_obc_sensors()
+        self.__publish_ccu_sensors()
+        self.__publish_location_sensors()
+        self.__publish_tyres_sensors()
+        # Binary sensors
+        self._publish_binary_sensor(
+            mqtt_topics.DRIVETRAIN_CHARGER_CONNECTED,
+            "Charger connected",
+            device_class="plug",
+            icon="mdi:power-plug-battery",
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DRIVETRAIN_HV_BATTERY_ACTIVE,
+            "HV Battery Active",
+            device_class="power",
+            icon="mdi:battery-check",
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DRIVETRAIN_CHARGING,
+            "Battery Charging",
+            device_class="battery_charging",
+            icon="mdi:battery-charging",
+        )
+        self._publish_sensor(
+            mqtt_topics.DRIVETRAIN_CHARGING_STOP_REASON,
+            "Battery charging stop reason",
+            icon="mdi:battery-charging",
+            enabled=False,
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DRIVETRAIN_BATTERY_HEATING,
+            "Battery heating",
+            icon="mdi:heat-wave",
+        )
+        self._publish_sensor(
+            mqtt_topics.DRIVETRAIN_BATTERY_HEATING_STOP_REASON,
+            "Battery heating stop reason",
+            icon="mdi:heat-wave",
+            enabled=False,
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DRIVETRAIN_RUNNING,
+            "Vehicle Running",
+            device_class="running",
+            icon="mdi:car-side",
+        )
+        self.__publish_lights_sensors()
+
+        LOG.debug("Completed publishing Home Assistant discovery messages")
+
+    def __publish_location_sensors(self) -> None:
+        self.__publish_vehicle_tracker()
+        self._publish_sensor(
+            mqtt_topics.LOCATION_HEADING,
+            "Heading",
+            icon="mdi:compass",
+            unit_of_measurement="°",
+        )
+        self._publish_sensor(
+            mqtt_topics.LOCATION_SPEED,
+            "Vehicle speed",
+            device_class="speed",
+            unit_of_measurement="km/h",
+        )
+        self._publish_switch(
+            mqtt_topics.LOCATION_FIND_MY_CAR,
+            "Find my car",
+            icon="mdi:car-search",
+            payload_on="activate",
+            payload_off="stop",
+        )
+
+    def __publish_windows_sensors(self) -> None:
+        self._publish_switch(mqtt_topics.WINDOWS_DRIVER, "Window driver")
+        self._publish_switch(mqtt_topics.WINDOWS_PASSENGER, "Window passenger")
+        self._publish_switch(mqtt_topics.WINDOWS_REAR_LEFT, "Window rear left")
+        self._publish_switch(mqtt_topics.WINDOWS_REAR_RIGHT, "Window rear right")
+        if self.__vin_info.has_sunroof:
+            self._publish_switch(mqtt_topics.WINDOWS_SUN_ROOF, "Sun roof")
+            self._publish_binary_sensor(mqtt_topics.WINDOWS_SUN_ROOF, "Sun roof")
+        else:
+            self.__unpublish_ha_discovery_message("switch", "Sun roof")
+            self.__unpublish_ha_discovery_message("binary_sensor", "Sun roof")
+
+    def __publish_ccu_sensors(self) -> None:
+        self._publish_sensor(
+            mqtt_topics.CCU_ONBOARD_PLUG_STATUS,
+            "CCU Onboard Plug Status",
+            state_class="measurement",
+            entity_category="diagnostic",
+            enabled=False,
+        )
+        self._publish_sensor(
+            mqtt_topics.CCU_OFFBOARD_PLUG_STATUS,
+            "CCU Offboard Plug Status",
+            state_class="measurement",
+            entity_category="diagnostic",
+            enabled=False,
+        )
+
+    def __publish_obc_sensors(self) -> None:
+        self._publish_sensor(
             mqtt_topics.OBC_CURRENT,
             "OBC Current",
             device_class="current",
@@ -540,7 +463,7 @@ class HomeAssistantDiscovery:
             entity_category="diagnostic",
             enabled=False,
         )
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.OBC_VOLTAGE,
             "OBC Voltage",
             device_class="voltage",
@@ -549,8 +472,7 @@ class HomeAssistantDiscovery:
             entity_category="diagnostic",
             enabled=False,
         )
-
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.OBC_POWER_SINGLE_PHASE,
             "OBC Power Single Phase",
             device_class="power",
@@ -559,8 +481,7 @@ class HomeAssistantDiscovery:
             entity_category="diagnostic",
             enabled=False,
         )
-
-        self.__publish_sensor(
+        self._publish_sensor(
             mqtt_topics.OBC_POWER_THREE_PHASE,
             "OBC Power Three Phase",
             device_class="power",
@@ -570,185 +491,239 @@ class HomeAssistantDiscovery:
             enabled=False,
         )
 
-        self.__publish_sensor(
-            mqtt_topics.CCU_ONBOARD_PLUG_STATUS,
-            "CCU Onboard Plug Status",
-            state_class="measurement",
-            entity_category="diagnostic",
-            enabled=False,
-        )
-
-        self.__publish_sensor(
-            mqtt_topics.CCU_OFFBOARD_PLUG_STATUS,
-            "CCU Offboard Plug Status",
-            state_class="measurement",
-            entity_category="diagnostic",
-            enabled=False,
-        )
-
-        self.__publish_sensor(
-            mqtt_topics.CLIMATE_INTERIOR_TEMPERATURE,
-            "Interior temperature",
-            device_class="temperature",
-            state_class="measurement",
-            unit_of_measurement="°C",
-        )
-        self.__publish_sensor(
-            mqtt_topics.CLIMATE_EXTERIOR_TEMPERATURE,
-            "Exterior temperature",
-            device_class="temperature",
-            state_class="measurement",
-            unit_of_measurement="°C",
-        )
-        self.__publish_sensor(
-            mqtt_topics.CLIMATE_REMOTE_CLIMATE_STATE,
-            "Remote climate state",
-            icon="mdi:car-connected",
-        )
-        self.__publish_sensor(
-            mqtt_topics.CLIMATE_BACK_WINDOW_HEAT,
-            "Rear window defroster heating",
-            icon="mdi:car-defrost-rear",
-        )
-        self.__publish_sensor(
-            mqtt_topics.LOCATION_HEADING,
-            "Heading",
-            icon="mdi:compass",
-            unit_of_measurement="°",
-        )
-        self.__publish_sensor(
-            mqtt_topics.LOCATION_SPEED,
-            "Vehicle speed",
-            device_class="speed",
-            unit_of_measurement="km/h",
-        )
-        self.__publish_sensor(
-            mqtt_topics.TYRES_FRONT_LEFT_PRESSURE,
-            "Tyres front left pressure",
-            device_class="pressure",
-            unit_of_measurement="bar",
-            icon="mdi:tire",
-        )
-        self.__publish_sensor(
-            mqtt_topics.TYRES_FRONT_RIGHT_PRESSURE,
-            "Tyres front right pressure",
-            device_class="pressure",
-            unit_of_measurement="bar",
-            icon="mdi:tire",
-        )
-        self.__publish_sensor(
-            mqtt_topics.TYRES_REAR_LEFT_PRESSURE,
-            "Tyres rear left pressure",
-            device_class="pressure",
-            unit_of_measurement="bar",
-            icon="mdi:tire",
-        )
-        self.__publish_sensor(
-            mqtt_topics.TYRES_REAR_RIGHT_PRESSURE,
-            "Tyres rear right pressure",
-            device_class="pressure",
-            unit_of_measurement="bar",
-            icon="mdi:tire",
-        )
-        # Binary sensors
-        self.__publish_binary_sensor(
-            mqtt_topics.DRIVETRAIN_CHARGER_CONNECTED,
-            "Charger connected",
-            device_class="plug",
-            icon="mdi:power-plug-battery",
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DRIVETRAIN_HV_BATTERY_ACTIVE,
-            "HV Battery Active",
-            device_class="power",
-            icon="mdi:battery-check",
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DRIVETRAIN_CHARGING,
-            "Battery Charging",
-            device_class="battery_charging",
-            icon="mdi:battery-charging",
-        )
-        self.__publish_sensor(
-            mqtt_topics.DRIVETRAIN_CHARGING_STOP_REASON,
-            "Battery charging stop reason",
-            icon="mdi:battery-charging",
-            enabled=False,
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DRIVETRAIN_BATTERY_HEATING,
-            "Battery heating",
-            icon="mdi:heat-wave",
-        )
-        self.__publish_sensor(
-            mqtt_topics.DRIVETRAIN_BATTERY_HEATING_STOP_REASON,
-            "Battery heating stop reason",
-            icon="mdi:heat-wave",
-            enabled=False,
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DRIVETRAIN_RUNNING,
-            "Vehicle Running",
-            device_class="running",
-            icon="mdi:car-side",
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DOORS_DRIVER,
-            "Door driver",
-            device_class="door",
-            icon="mdi:car-door",
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DOORS_PASSENGER,
-            "Door passenger",
-            device_class="door",
-            icon="mdi:car-door",
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DOORS_REAR_LEFT,
-            "Door rear left",
-            device_class="door",
-            icon="mdi:car-door",
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DOORS_REAR_RIGHT,
-            "Door rear right",
-            device_class="door",
-            icon="mdi:car-door",
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DOORS_BONNET, "Bonnet", device_class="door", icon="mdi:car-door"
-        )
-        self.__publish_binary_sensor(
-            mqtt_topics.DOORS_BOOT, "Boot", device_class="door", icon="mdi:car-door"
-        )
-        self.__publish_binary_sensor(
+    def __publish_lights_sensors(self) -> None:
+        self._publish_binary_sensor(
             mqtt_topics.LIGHTS_MAIN_BEAM,
             "Lights Main Beam",
             device_class="light",
             icon="mdi:car-light-high",
         )
-        self.__publish_binary_sensor(
+        self._publish_binary_sensor(
             mqtt_topics.LIGHTS_DIPPED_BEAM,
             "Lights Dipped Beam",
             device_class="light",
             icon="mdi:car-light-dimmed",
         )
-        self.__publish_binary_sensor(
+        self._publish_binary_sensor(
             mqtt_topics.LIGHTS_SIDE,
             "Lights Side",
             device_class="light",
             icon="mdi:car-light-dimmed",
         )
 
+    def __publish_tyres_sensors(self) -> None:
+        self._publish_sensor(
+            mqtt_topics.TYRES_FRONT_LEFT_PRESSURE,
+            "Tyres front left pressure",
+            device_class="pressure",
+            unit_of_measurement="bar",
+            icon="mdi:tire",
+        )
+        self._publish_sensor(
+            mqtt_topics.TYRES_FRONT_RIGHT_PRESSURE,
+            "Tyres front right pressure",
+            device_class="pressure",
+            unit_of_measurement="bar",
+            icon="mdi:tire",
+        )
+        self._publish_sensor(
+            mqtt_topics.TYRES_REAR_LEFT_PRESSURE,
+            "Tyres rear left pressure",
+            device_class="pressure",
+            unit_of_measurement="bar",
+            icon="mdi:tire",
+        )
+        self._publish_sensor(
+            mqtt_topics.TYRES_REAR_RIGHT_PRESSURE,
+            "Tyres rear right pressure",
+            device_class="pressure",
+            unit_of_measurement="bar",
+            icon="mdi:tire",
+        )
+
+    def __publish_doors_sensors(self) -> None:
+        self._publish_lock(
+            mqtt_topics.DOORS_LOCKED, "Doors Lock", icon="mdi:car-door-lock"
+        )
+        self._publish_lock(
+            mqtt_topics.DOORS_BOOT,
+            "Boot Lock",
+            icon="mdi:car-door-lock",
+            state_locked="False",
+            state_unlocked="True",
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DOORS_DRIVER,
+            "Door driver",
+            device_class="door",
+            icon="mdi:car-door",
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DOORS_PASSENGER,
+            "Door passenger",
+            device_class="door",
+            icon="mdi:car-door",
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DOORS_REAR_LEFT,
+            "Door rear left",
+            device_class="door",
+            icon="mdi:car-door",
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DOORS_REAR_RIGHT,
+            "Door rear right",
+            device_class="door",
+            icon="mdi:car-door",
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DOORS_BONNET, "Bonnet", device_class="door", icon="mdi:car-door"
+        )
+        self._publish_binary_sensor(
+            mqtt_topics.DOORS_BOOT, "Boot", device_class="door", icon="mdi:car-door"
+        )
+
+    def __publish_gateway_sensors(self) -> None:
+        options = [m.value for m in RefreshMode]
+        self._publish_select(
+            mqtt_topics.REFRESH_MODE,
+            "Gateway refresh mode",
+            options,
+            entity_category="config",
+            enabled=True,
+            value_template="{{ value }}",
+            command_template="{{ value }}",
+            icon="mdi:refresh",
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_number(
+            mqtt_topics.REFRESH_PERIOD_ACTIVE,
+            "Gateway active refresh period",
+            entity_category="config",
+            unit_of_measurement="s",
+            icon="mdi:timer",
+            min_value=30,
+            max_value=60 * 60,
+            step=1,
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_number(
+            mqtt_topics.REFRESH_PERIOD_INACTIVE,
+            "Gateway inactive refresh period",
+            entity_category="config",
+            unit_of_measurement="s",
+            icon="mdi:timer",
+            min_value=1 * 60 * 60,
+            max_value=5 * 24 * 60 * 60,
+            step=1,
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_number(
+            mqtt_topics.REFRESH_PERIOD_AFTER_SHUTDOWN,
+            "Gateway refresh period after car shutdown",
+            entity_category="config",
+            unit_of_measurement="s",
+            icon="mdi:timer",
+            min_value=30,
+            max_value=12 * 60 * 60,
+            step=1,
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_number(
+            mqtt_topics.REFRESH_PERIOD_INACTIVE_GRACE,
+            "Gateway grace period after car shutdown",
+            entity_category="config",
+            unit_of_measurement="s",
+            icon="mdi:timer",
+            min_value=30,
+            max_value=12 * 60 * 60,
+            step=1,
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_sensor(
+            mqtt_topics.REFRESH_PERIOD_CHARGING,
+            "Gateway charging refresh period",
+            entity_category="diagnostic",
+            unit_of_measurement="s",
+            icon="mdi:timer",
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_sensor(
+            mqtt_topics.REFRESH_PERIOD_ERROR,
+            "Gateway error refresh period",
+            entity_category="diagnostic",
+            unit_of_measurement="s",
+            icon="mdi:timer",
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_sensor(
+            mqtt_topics.REFRESH_LAST_ACTIVITY,
+            "Last car activity",
+            device_class="timestamp",
+            entity_category="diagnostic",
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_sensor(
+            mqtt_topics.REFRESH_LAST_CHARGE_STATE,
+            "Last charge state",
+            device_class="timestamp",
+            entity_category="diagnostic",
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_sensor(
+            mqtt_topics.REFRESH_LAST_VEHICLE_STATE,
+            "Last vehicle state",
+            device_class="timestamp",
+            entity_category="diagnostic",
+            custom_availability=self.__system_availability_config,
+        )
+        self._publish_sensor(
+            mqtt_topics.REFRESH_LAST_ERROR,
+            "Last poll error",
+            device_class="timestamp",
+            entity_category="diagnostic",
+            custom_availability=self.__system_availability_config,
+        )
+
+    def __publish_climate_sensors(self) -> None:
+        self.__publish_remote_ac()
+        self.__publish_heated_seats()
+        self.__publish_back_window_heat()
+        self._publish_sensor(
+            mqtt_topics.CLIMATE_INTERIOR_TEMPERATURE,
+            "Interior temperature",
+            device_class="temperature",
+            state_class="measurement",
+            unit_of_measurement="°C",
+        )
+        self._publish_sensor(
+            mqtt_topics.CLIMATE_EXTERIOR_TEMPERATURE,
+            "Exterior temperature",
+            device_class="temperature",
+            state_class="measurement",
+            unit_of_measurement="°C",
+        )
         # Remove deprecated sensors
         self.__unpublish_ha_discovery_message(
             "sensor", "Front window defroster heating"
         )
-        LOG.debug("Completed publishing Home Assistant discovery messages")
+
+    def __publish_back_window_heat(self) -> None:
+        self._publish_switch(
+            mqtt_topics.CLIMATE_BACK_WINDOW_HEAT,
+            "Rear window defroster heating",
+            icon="mdi:car-defrost-rear",
+            payload_on="on",
+            payload_off="off",
+        )
+        self._publish_sensor(
+            mqtt_topics.CLIMATE_BACK_WINDOW_HEAT,
+            "Rear window defroster heating",
+            icon="mdi:car-defrost-rear",
+        )
 
     def __publish_vehicle_tracker(self) -> None:
-        self.__publish_ha_discovery_message(
+        self._publish_ha_discovery_message(
             "device_tracker",
             "Vehicle position",
             {
@@ -761,7 +736,13 @@ class HomeAssistantDiscovery:
     def __publish_remote_ac(self) -> None:
         # This has been converted into 2 switches and a climate entity for ease of operation
 
-        self.__publish_switch(
+        self._publish_sensor(
+            mqtt_topics.CLIMATE_REMOTE_CLIMATE_STATE,
+            "Remote climate state",
+            icon="mdi:car-connected",
+        )
+
+        self._publish_switch(
             mqtt_topics.CLIMATE_REMOTE_CLIMATE_STATE,
             "Front window defroster heating",
             icon="mdi:car-defrost-front",
@@ -770,7 +751,7 @@ class HomeAssistantDiscovery:
             payload_off="off",
         )
 
-        self.__publish_switch(
+        self._publish_switch(
             mqtt_topics.CLIMATE_REMOTE_CLIMATE_STATE,
             "Vehicle climate fan only",
             icon="mdi:fan",
@@ -779,7 +760,7 @@ class HomeAssistantDiscovery:
             payload_off="off",
         )
 
-        self.__publish_ha_discovery_message(
+        self._publish_ha_discovery_message(
             "climate",
             "Vehicle climate",
             {
@@ -809,233 +790,6 @@ class HomeAssistantDiscovery:
                 "min_temp": self.__vin_info.min_ac_temperature,
                 "max_temp": self.__vin_info.max_ac_temperature,
             },
-        )
-
-    def __publish_switch(
-        self,
-        topic: str,
-        name: str,
-        *,
-        enabled: bool = True,
-        icon: str | None = None,
-        value_template: str = "{{ value }}",
-        payload_on: str = "True",
-        payload_off: str = "False",
-        custom_availability: HaCustomAvailabilityConfig | None = None,
-    ) -> str:
-        payload = {
-            "state_topic": self.__get_vehicle_topic(topic),
-            "command_topic": self.__get_vehicle_set_topic(topic),
-            "value_template": value_template,
-            "payload_on": payload_on,
-            "payload_off": payload_off,
-            "optimistic": False,
-            "qos": 0,
-            "enabled_by_default": enabled,
-        }
-        if icon is not None:
-            payload["icon"] = icon
-        return self.__publish_ha_discovery_message(
-            "switch", name, payload, custom_availability
-        )
-
-    def __publish_lock(
-        self,
-        topic: str,
-        name: str,
-        enabled: bool = True,
-        icon: str | None = None,
-        payload_lock: str = "True",
-        payload_unlock: str = "False",
-        state_locked: str = "True",
-        state_unlocked: str = "False",
-        custom_availability: HaCustomAvailabilityConfig | None = None,
-    ) -> str:
-        payload = {
-            "state_topic": self.__get_vehicle_topic(topic),
-            "command_topic": self.__get_vehicle_set_topic(topic),
-            "payload_lock": payload_lock,
-            "payload_unlock": payload_unlock,
-            "state_locked": state_locked,
-            "state_unlocked": state_unlocked,
-            "optimistic": False,
-            "qos": 0,
-            "enabled_by_default": enabled,
-        }
-        if icon is not None:
-            payload["icon"] = icon
-        return self.__publish_ha_discovery_message(
-            "lock", name, payload, custom_availability
-        )
-
-    def __publish_sensor(
-        self,
-        topic: str,
-        name: str,
-        enabled: bool = True,
-        entity_category: str | None = None,
-        device_class: str | None = None,
-        state_class: str | None = None,
-        unit_of_measurement: str | None = None,
-        icon: str | None = None,
-        value_template: str = "{{ value }}",
-        custom_availability: HaCustomAvailabilityConfig | None = None,
-    ) -> str:
-        payload = {
-            "state_topic": self.__get_vehicle_topic(topic),
-            "value_template": value_template,
-            "enabled_by_default": enabled,
-        }
-        if entity_category is not None:
-            payload["entity_category"] = entity_category
-        if device_class is not None:
-            payload["device_class"] = device_class
-        if state_class is not None:
-            payload["state_class"] = state_class
-        if unit_of_measurement is not None:
-            payload["unit_of_measurement"] = unit_of_measurement
-        if icon is not None:
-            payload["icon"] = icon
-
-        return self.__publish_ha_discovery_message(
-            "sensor", name, payload, custom_availability
-        )
-
-    def __publish_number(
-        self,
-        topic: str,
-        name: str,
-        enabled: bool = True,
-        entity_category: str | None = None,
-        device_class: str | None = None,
-        state_class: str | None = None,
-        unit_of_measurement: str | None = None,
-        icon: str | None = None,
-        value_template: str = "{{ value }}",
-        retain: bool = False,
-        mode: str = "auto",
-        min_value: float = 1.0,
-        max_value: float = 100.0,
-        step: float = 1.0,
-        custom_availability: HaCustomAvailabilityConfig | None = None,
-    ) -> str:
-        payload = {
-            "state_topic": self.__get_vehicle_topic(topic),
-            "command_topic": self.__get_vehicle_set_topic(topic),
-            "value_template": value_template,
-            "retain": str(retain).lower(),
-            "mode": mode,
-            "min": min_value,
-            "max": max_value,
-            "step": step,
-            "enabled_by_default": enabled,
-        }
-        if entity_category is not None:
-            payload["entity_category"] = entity_category
-        if device_class is not None:
-            payload["device_class"] = device_class
-        if state_class is not None:
-            payload["state_class"] = state_class
-        if unit_of_measurement is not None:
-            payload["unit_of_measurement"] = unit_of_measurement
-        if icon is not None:
-            payload["icon"] = icon
-
-        return self.__publish_ha_discovery_message(
-            "number", name, payload, custom_availability
-        )
-
-    def __publish_text(
-        self,
-        topic: str,
-        name: str,
-        enabled: bool = True,
-        icon: str | None = None,
-        value_template: str = "{{ value }}",
-        command_template: str = "{{ value }}",
-        retain: bool = False,
-        min_value: int | None = None,
-        max_value: int | None = None,
-        pattern: str | None = None,
-        custom_availability: HaCustomAvailabilityConfig | None = None,
-    ) -> str:
-        payload = {
-            "state_topic": self.__get_vehicle_topic(topic),
-            "command_topic": self.__get_vehicle_set_topic(topic),
-            "value_template": value_template,
-            "command_template": command_template,
-            "retain": str(retain).lower(),
-            "enabled_by_default": enabled,
-        }
-        if min_value is not None:
-            payload["min"] = min_value
-        if max_value is not None:
-            payload["max"] = max_value
-        if pattern is not None:
-            payload["pattern"] = pattern
-        if icon is not None:
-            payload["icon"] = icon
-
-        return self.__publish_ha_discovery_message(
-            "text", name, payload, custom_availability
-        )
-
-    def __publish_binary_sensor(
-        self,
-        topic: str,
-        name: str,
-        enabled: bool = True,
-        device_class: str | None = None,
-        value_template: str = "{{ value }}",
-        payload_on: str = "True",
-        payload_off: str = "False",
-        icon: str | None = None,
-        custom_availability: HaCustomAvailabilityConfig | None = None,
-    ) -> str:
-        payload = {
-            "state_topic": self.__get_vehicle_topic(topic),
-            "value_template": value_template,
-            "payload_on": payload_on,
-            "payload_off": payload_off,
-            "enabled_by_default": enabled,
-        }
-        if device_class is not None:
-            payload["device_class"] = device_class
-        if icon is not None:
-            payload["icon"] = icon
-
-        return self.__publish_ha_discovery_message(
-            "binary_sensor", name, payload, custom_availability
-        )
-
-    def __publish_select(
-        self,
-        topic: str,
-        name: str,
-        options: list[str],
-        *,
-        entity_category: str | None = None,
-        enabled: bool = True,
-        value_template: str = "{{ value }}",
-        command_template: str = "{{ value }}",
-        icon: str | None = None,
-        custom_availability: HaCustomAvailabilityConfig | None = None,
-    ) -> str:
-        payload = {
-            "state_topic": self.__get_vehicle_topic(topic),
-            "command_topic": self.__get_vehicle_set_topic(topic),
-            "value_template": value_template,
-            "command_template": command_template,
-            "options": options,
-            "enabled_by_default": enabled,
-        }
-        if entity_category is not None:
-            payload["entity_category"] = entity_category
-        if icon is not None:
-            payload["icon"] = icon
-
-        return self.__publish_ha_discovery_message(
-            "select", name, payload, custom_availability
         )
 
     def __get_common_attributes(
@@ -1087,6 +841,10 @@ class HomeAssistantDiscovery:
             return publisher.get_topic(topic, no_prefix=False)
         return topic
 
+    @override
+    def _get_state_topic(self, raw_topic: str) -> str:
+        return self.__get_vehicle_topic(raw_topic)
+
     def __get_vehicle_topic(self, topic: str) -> str:
         vehicle_topic: str = self.__vehicle_state.get_topic(topic)
         publisher = self.__vehicle_state.publisher
@@ -1094,10 +852,15 @@ class HomeAssistantDiscovery:
             return publisher.get_topic(vehicle_topic, no_prefix=False)
         return vehicle_topic
 
+    @override
+    def _get_command_topic(self, raw_topic: str) -> str:
+        return self.__get_vehicle_set_topic(raw_topic)
+
     def __get_vehicle_set_topic(self, topic: str) -> str:
         return self.__get_vehicle_topic(topic) + "/" + mqtt_topics.SET_SUFFIX
 
-    def __publish_ha_discovery_message(
+    @override
+    def _publish_ha_discovery_message(
         self,
         sensor_type: str,
         sensor_name: str,
@@ -1130,19 +893,19 @@ class HomeAssistantDiscovery:
         self.__vehicle_state.publisher.publish_str(ha_topic, "", no_prefix=True)
 
     def __publish_scheduled_charging(self) -> None:
-        start_time_id = self.__publish_sensor(
+        start_time_id = self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CHARGING_SCHEDULE,
             "Scheduled Charging Start",
             value_template='{{ value_json["startTime"] }}',
             icon="mdi:clock-start",
         )
-        end_time_id = self.__publish_sensor(
+        end_time_id = self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CHARGING_SCHEDULE,
             "Scheduled Charging End",
             value_template='{{ value_json["endTime"] }}',
             icon="mdi:clock-end",
         )
-        scheduled_charging_mode_id = self.__publish_sensor(
+        scheduled_charging_mode_id = self._publish_sensor(
             mqtt_topics.DRIVETRAIN_CHARGING_SCHEDULE,
             "Scheduled Charging Mode",
             value_template='{{ value_json["mode"] }}',
@@ -1156,13 +919,17 @@ class HomeAssistantDiscovery:
                 "mode": "{{ value }}",
             }
         )
-        self.__publish_select(
+        options = [m.name for m in ScheduledChargingMode]
+        self._publish_select(
             mqtt_topics.DRIVETRAIN_CHARGING_SCHEDULE,
             "Scheduled Charging Mode",
-            options=[m.name for m in ScheduledChargingMode],
+            options,
+            entity_category=None,
+            enabled=True,
             value_template='{{ value_json["mode"] }}',
             command_template=change_mode_cmd_template,
             icon="mdi:clock-outline",
+            custom_availability=None,
         )
 
         change_start_cmd_template = json.dumps(
@@ -1172,15 +939,15 @@ class HomeAssistantDiscovery:
                 "mode": f"{{{{ states('{scheduled_charging_mode_id}') }}}}",
             }
         )
-        self.__publish_text(
+        self._publish_text(
             mqtt_topics.DRIVETRAIN_CHARGING_SCHEDULE,
             "Scheduled Charging Start",
+            icon="mdi:clock-start",
             value_template='{{ value_json["startTime"] }}',
             command_template=change_start_cmd_template,
             min_value=4,
             max_value=5,
             pattern="^([01][0-9]|2[0-3]):[0-5][0-9]$",
-            icon="mdi:clock-start",
         )
 
         change_end_cmd_template = json.dumps(
@@ -1190,56 +957,60 @@ class HomeAssistantDiscovery:
                 "mode": f"{{{{ states('{scheduled_charging_mode_id}') }}}}",
             }
         )
-        self.__publish_text(
+        self._publish_text(
             mqtt_topics.DRIVETRAIN_CHARGING_SCHEDULE,
             "Scheduled Charging End",
+            icon="mdi:clock-end",
             value_template='{{ value_json["endTime"] }}',
             command_template=change_end_cmd_template,
             min_value=4,
             max_value=5,
             pattern="^([01][0-9]|2[0-3]):[0-5][0-9]$",
-            icon="mdi:clock-end",
         )
 
     def __publish_scheduled_battery_heating(self) -> None:
-        start_time_id = self.__publish_sensor(
+        start_time_id = self._publish_sensor(
             mqtt_topics.DRIVETRAIN_BATTERY_HEATING_SCHEDULE,
             "Scheduled Battery Heating Start",
             value_template='{{ value_json["startTime"] }}',
             icon="mdi:clock-start",
         )
-        mode_id = self.__publish_binary_sensor(
+        mode_id = self._publish_binary_sensor(
             mqtt_topics.DRIVETRAIN_BATTERY_HEATING_SCHEDULE,
             "Scheduled Battery Heating",
             value_template='{{ value_json["mode"] }}',
-            icon="mdi:clock-outline",
             payload_on="on",
             payload_off="off",
+            icon="mdi:clock-outline",
         )
         change_mode_cmd_template = json.dumps(
             {"startTime": f"{{{{ states('{start_time_id}') }}}}", "mode": "{{ value }}"}
         )
-        self.__publish_select(
+        options = ["on", "off"]
+        self._publish_select(
             mqtt_topics.DRIVETRAIN_BATTERY_HEATING_SCHEDULE,
             "Scheduled Battery Heating",
-            options=["on", "off"],
+            options,
+            entity_category=None,
+            enabled=True,
             value_template='{{ value_json["mode"] }}',
             command_template=change_mode_cmd_template,
             icon="mdi:clock-outline",
+            custom_availability=None,
         )
 
         change_start_cmd_template = json.dumps(
             {"startTime": "{{ value }}", "mode": f"{{{{ states('{mode_id}') }}}}"}
         )
-        self.__publish_text(
+        self._publish_text(
             mqtt_topics.DRIVETRAIN_BATTERY_HEATING_SCHEDULE,
             "Scheduled Battery Heating Start",
+            icon="mdi:clock-start",
             value_template='{{ value_json["startTime"] }}',
             command_template=change_start_cmd_template,
             min_value=4,
             max_value=5,
             pattern="^([01][0-9]|2[0-3]):[0-5][0-9]$",
-            icon="mdi:clock-start",
         )
 
     def __publish_heated_seats(self) -> None:
@@ -1268,10 +1039,13 @@ class HomeAssistantDiscovery:
             self.__unpublish_heated_seat_switch("Front Right")
 
     def __publish_heated_seat_level(self, seat: str, topic: str) -> None:
-        self.__publish_select(
+        options = ["OFF", "LOW", "MEDIUM", "HIGH"]
+        self._publish_select(
             topic,
             f"Heated Seat {seat} Level",
-            options=["OFF", "LOW", "MEDIUM", "HIGH"],
+            options,
+            entity_category=None,
+            enabled=True,
             value_template="{% set v = value | int %}"
             "{% if v == 0 %}OFF"
             "{% elif v == 1 %}LOW"
@@ -1284,44 +1058,22 @@ class HomeAssistantDiscovery:
             "{% else %}3"
             "{% endif %}",
             icon="mdi:car-seat-heater",
+            custom_availability=None,
         )
 
     def __unpublish_heated_seat_level(self, seat: str) -> None:
         self.__unpublish_ha_discovery_message("select", f"Heated Seat {seat} Level")
 
     def __publish_heated_seat_switch(self, seat: str, topic: str) -> None:
-        self.__publish_switch(
+        self._publish_switch(
             topic,
             f"Heated Seat {seat}",
-            payload_off="0",
-            payload_on="1",
             icon="mdi:car-seat-heater",
+            payload_on="1",
+            payload_off="0",
         )
 
     def __unpublish_heated_seat_switch(self, seat: str) -> None:
         self.__unpublish_ha_discovery_message("switch", f"Heated Seat {seat}")
 
 
-def snake_case(s: str) -> str:
-    return inflection.underscore(s.lower()).replace(" ", "_")
-
-
-def decode_as_utf8(
-    byte_string: str | None | bytes | bytearray, default: str = ""
-) -> str:
-    if byte_string is None:
-        return default
-    if isinstance(byte_string, str):
-        return byte_string
-    if isinstance(byte_string, bytes | bytearray):
-        try:
-            return str(byte_string, encoding="utf8", errors="ignore")
-        except Exception:
-            LOG.exception(f"Failed to decode {byte_string!r} as utf8")
-            return default
-    else:
-        try:  # type: ignore[unreachable]
-            return str(byte_string)
-        except Exception:
-            LOG.exception(f"Failed to decode {byte_string}")
-            return default
